Completed bulk batch. Saving settings before bulk synchronization. Added debug logs. Added specific permission management for Facebook Shop synchronization.

+++ src/Modules/Laser.Orchard.NwazetIntegration/Handlers/FacebookShopProductPartHandler.cs
@@ -1,26 +1,31 @@
 ﻿using Laser.Orchard.NwazetIntegration.Models;
+using Laser.Orchard.NwazetIntegration.Security;
 using Laser.Orchard.NwazetIntegration.Services.FacebookShop;
 using Orchard.ContentManagement;
 using Orchard.ContentManagement.Handlers;
 using Orchard.Data;
 using Orchard.Environment.Extensions;
 using Orchard.Localization;
+using Orchard.Security;
 using Orchard.UI.Notify;
 
 namespace Laser.Orchard.NwazetIntegration.Handlers {
     [OrchardFeature("Laser.Orchard.FacebookShop")]
     public class FacebookShopProductPartHandler : ContentHandler {
-        IFacebookShopService _facebookShopService;
-        INotifier _notifier;
+        private readonly IFacebookShopService _facebookShopService;
+        private readonly INotifier _notifier;
+        private readonly IAuthorizer _authorizer;
 
         public FacebookShopProductPartHandler(
             IRepository<FacebookShopProductPartRecord> repository,
             IFacebookShopService facebookShopService,
-            INotifier notifier) {
+            INotifier notifier,
+            IAuthorizer authorizer) {
 
             Filters.Add(StorageFilter.For(repository));
             _facebookShopService = facebookShopService;
             _notifier = notifier;
+            _authorizer = authorizer;
 
             T = NullLocalizer.Instance;
 
@@ -34,12 +39,14 @@ namespace Laser.Orchard.NwazetIntegration.Handlers {
         public Localizer T { get; set; }
 
         protected void AfterPublish(PublishContentContext context) {
-            var result = _facebookShopService.SyncProduct(context.ContentItem);
+            if (_authorizer.Authorize(FacebookShopSynchronizationPermission.FacebookShopSynchronization)) {
+                var result = _facebookShopService.SyncProduct(context.ContentItem);
 
-            if (result != null && result.Valid) {
-
-            } else {
-                _notifier.Warning(result.Message);
+                if (result == null || result.Requests.Count == 0) {
+                    _notifier.Warning(T("Invalid Facebook api response. Product is not synchronised on Facebook catalog."));
+                } else if (!result.Requests[0].Valid) {
+                    _notifier.Warning(result.Requests[0].Message);
+                }
             }
         }
 
@@ -52,12 +59,14 @@ namespace Laser.Orchard.NwazetIntegration.Handlers {
         }
 
         private void RemoveProduct(ContentItem product) {
-            var result = _facebookShopService.RemoveProduct(product);
+            if (_authorizer.Authorize(FacebookShopSynchronizationPermission.FacebookShopSynchronization)) {
+                var result = _facebookShopService.RemoveProduct(product);
 
-            if (result != null && result.Valid) {
-
-            } else {
-                _notifier.Warning(result.Message);
+                if (result == null || result.Requests.Count == 0) {
+                    _notifier.Warning(T("Invalid Facebook api response. Product has not been removed from Facebook catalog."));
+                } else if (!result.Requests[0].Valid) {
+                    _notifier.Warning(result.Requests[0].Message);
+                }
             }
         }
     }
